Fature flag removal of TaskSearch improvement  (#1165)
* removal of feature flag

* fix TaskContent test

* remove TaskSearch deprecated component

* fix test for removing TaskSearch deprecated component

* remove TabSection component

* rename TaskSearchDev to TaskSearch

* rename TaskSearchDev to TaskSearch in tests

* resolve merge conflict
