+++ src/components/tasks/TaskSearch/FilterDropdown.tsx
@@ -1,9 +1,5 @@
 import styles from './tasksearch.module.scss';
-import {
-    Tab,
-    depreciatedTaskStatus,
-    newTaskStatus,
-} from '@/interfaces/task.type';
+import { Tab, depreciatedTaskStatus } from '@/interfaces/task.type';
 import { getChangedStatusName } from '@/utils/getChangedStatusName';
 import { useEffect } from 'react';
 
@@ -12,18 +8,16 @@ type FilterModalProps = {
     onSelect: (tab: Tab) => void;
     activeTab?: Tab;
     onClose: () => void;
-    dev?: boolean;
 };
 
-const FilterModal = ({
+const FilterDropdown = ({
     tabs,
     onSelect,
     activeTab,
     onClose,
-    dev,
 }: FilterModalProps) => {
     const onKeyDownHandler = (event: KeyboardEvent) => {
-        if (dev && event.key === 'Escape') onClose();
+        onClose();
     };
     useEffect(() => {
         document.addEventListener('keydown', onKeyDownHandler);
@@ -32,16 +26,12 @@ const FilterModal = ({
 
     return (
         <>
-            {dev && (
-                <div
-                    onClick={onClose}
-                    className={styles['filter-modal-background']}
-                ></div>
-            )}
             <div
-                className={`${styles['filter-modal']} ${
-                    dev ? styles['filter-modal-dev'] : ''
-                }`}
+                onClick={onClose}
+                className={styles['filter-modal-background']}
+            ></div>
+            <div
+                className={`${styles['filter-modal']} ${styles['filter-modal-dev']}`}
                 data-testid="filter-modal"
             >
                 <div className={styles['filter-modal-title']}>
@@ -52,11 +42,8 @@ const FilterModal = ({
                 </div>
                 <div className={styles['status-filter']}>
                     {tabs
-                        .filter((tab: Tab) =>
-                            dev
-                                ? !depreciatedTaskStatus.includes(tab)
-                                : tab != 'BACKLOG' &&
-                                  !newTaskStatus.includes(tab)
+                        .filter(
+                            (tab: Tab) => !depreciatedTaskStatus.includes(tab)
                         )
                         .map((tab) => (
                             <button
@@ -80,4 +67,4 @@ const FilterModal = ({
     );
 };
 
-export default FilterModal;
+export default FilterDropdown;
