+++ src/components/tasks/TaskSearch/tasksearch.module.scss
@@ -37,6 +37,7 @@
 }
 
 .filter-button {
+    width: 5rem;
     height: 2rem;
     position: relative;
     border-radius: $radius 0 0 $radius;
@@ -50,17 +51,19 @@
     display: flex;
     flex-direction: row;
     align-items: center;
-    column-gap: 0.4rem;
-    padding: 0 0.5rem;
-}
+    padding-left: 5px;
 
-.filter-chevron {
-    width: 1rem;
-    height: 1rem;
-}
+    &:after {
+        content: '';
+        border-left: 5px solid transparent;
+        border-right: 5px solid transparent;
+        border-top: 5px solid $black;
+        margin-left: 0.5rem;
+    }
 
-.filter-chevron-open {
-    transform: rotate(180deg);
+    @media (max-width: 768px) {
+        width: 6rem;
+    }
 }
 
 .search-bar-div {
@@ -213,9 +216,9 @@
 }
 
 .filter-modal {
-    position: absolute;
-    top: 2.534rem;
-    left: 0;
+    position: relative;
+    top: 11.5rem;
+    left: -2.8rem;
     min-width: 15rem;
     background-color: $white;
     border-radius: $radius;
@@ -225,6 +228,13 @@
     justify-content: space-between;
     align-items: center;
     z-index: 3;
+    @media (max-width: 768px) {
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        width: 80%;
+    }
 }
 .filter-modal-dev {
     top: 13.5rem;
