--- src/components/tasks/TaskSearchDev/tasksearch.module.scss
@@ -1,301 +1,0 @@
-@import '../../../styles/variables.scss';
-
-.task-search-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
-
-    @media (max-width: 768px) {
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-    }
-}
-
-.filter-container {
-    width: 100%;
-    display: flex;
-    position: relative;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    border-radius: $radius;
-    margin-right: 1rem;
-    border: 1px solid $lighter-gray;
-
-    @media (max-width: 768px) {
-        width: 100%;
-        margin-right: 0;
-    }
-
-    @media (max-width: 368px) {
-        width: 90%;
-        margin-right: 0;
-    }
-}
-
-.filter-button {
-    width: 5rem;
-    height: 2rem;
-    position: relative;
-    border-radius: $radius 0 0 $radius;
-    background-color: $white;
-    color: $black;
-    font-weight: bold;
-    cursor: pointer;
-    border: none;
-    border-right: none;
-    outline: none;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding-left: 5px;
-
-    &:after {
-        content: '';
-        border-left: 5px solid transparent;
-        border-right: 5px solid transparent;
-        border-top: 5px solid $black;
-        margin-left: 0.5rem;
-    }
-
-    @media (max-width: 768px) {
-        width: 6rem;
-    }
-}
-
-.search-bar-div {
-    width: 43vw;
-    flex: 1;
-    min-height: 2rem;
-    border-radius: 0 $radius $radius 0;
-    background-color: $white;
-    border-left: 1px solid $lighter-gray;
-    background-image: url('https://img.icons8.com/ios-filled/16/000000/search.png');
-    background-repeat: no-repeat;
-    background-position: 0.5rem;
-    padding-left: 1.8rem;
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-}
-.in-focus {
-    border: 1px solid $gray;
-    cursor: pointer;
-    color: $black;
-}
-.search-input {
-    width: 100%;
-    height: 2rem;
-    padding: 0.5rem 0.5rem 0.5rem 2rem;
-    border-radius: 0 $radius $radius 0;
-    border: none;
-    outline: none;
-    background-image: url('https://img.icons8.com/ios-filled/16/000000/search.png');
-    background-repeat: no-repeat;
-    background-position: 0.5rem;
-    padding-left: 2rem;
-    background-color: $white;
-    border-left: 1px solid $lighter-gray;
-
-    &:focus {
-        border: 1px solid $theme-primary;
-    }
-}
-.search-bar-div:focus-within {
-    border: 1px solid $gray;
-    cursor: pointer;
-    color: $black;
-    margin-top: -1px;
-    margin-bottom: -1px;
-}
-.search-input-parent {
-    display: flex;
-    flex-wrap: wrap;
-    padding-top: 4px;
-    margin-top: -3px;
-
-    @media (min-width: 558px) {
-        min-width: 347px;
-    }
-    @media (max-width: 557px) {
-        width: 100% !important;
-    }
-    @media (max-width: 562px) and (min-width: 558px) {
-        min-width: 302px !important;
-    }
-    @media (max-width: 617px) and (min-width: 601px) {
-        min-width: 310px !important;
-    }
-    @media (max-width: 793px) and (min-width: 769px) {
-        min-width: 328px !important;
-    }
-}
-.pill-input-wrapper {
-    width: 100%;
-    padding: 3px 5px 3px 5px;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    gap: 2px;
-}
-.remove-caret {
-    caret-color: transparent;
-}
-.search-input-dev {
-    width: 100%;
-    height: 27px;
-    padding: 0.45rem;
-    border-radius: 0 $radius $radius 0;
-    border: none;
-    outline: none;
-}
-.suggestion-box {
-    width: 100%;
-    top: 104%;
-    z-index: 5;
-    position: absolute;
-    text-align: left;
-    background-color: $white;
-}
-.empty-box {
-    padding: 5.3px;
-    text-align: center;
-    border: 1px solid $lighter-gray;
-    border-top: none;
-}
-
-.search-button-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    width: 10rem;
-    @media (max-width: 768px) {
-        width: 100%;
-        margin-top: 1rem;
-    }
-    @media (max-width: 368px) {
-        width: 90%;
-        margin-top: 1rem;
-    }
-}
-
-.search-button {
-    width: 100%;
-    height: 2rem;
-    border-radius: $radius;
-    background-color: $white;
-    color: $black;
-    font-weight: bold;
-    cursor: pointer;
-    border: 1px solid $gray;
-    outline: none;
-
-    &:hover {
-        background-color: $theme-primary;
-        color: $white;
-        border: 1px solid $theme-primary;
-    }
-
-    @media (max-width: 768px) {
-        width: 100%;
-        border-radius: $radius;
-        margin-left: 0;
-    }
-}
-.filter-modal-background {
-    position: fixed;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 3;
-}
-
-.filter-modal {
-    position: relative;
-    top: 11.5rem;
-    left: -2.8rem;
-    min-width: 15rem;
-    background-color: $white;
-    border-radius: $radius;
-    border: 1px solid $light-gray;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    z-index: 3;
-    @media (max-width: 768px) {
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        width: 80%;
-    }
-}
-.filter-modal-dev {
-    top: 13.5rem;
-
-    @media (max-width: 768px) {
-        top: 14.5rem;
-        left: 7.5rem;
-    }
-}
-
-.filter-modal-title {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    border-bottom: 1px solid $gray;
-    padding: 0.5rem 0.5rem 0.5rem 1rem;
-
-    span {
-        font-size: 1rem;
-        font-weight: bold;
-    }
-}
-
-.status-filter {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    @media (max-width: 768px) {
-        left: 0;
-        top: 0;
-        width: 100%;
-        height: 100%;
-    }
-
-    .status-button {
-        width: 100%;
-        height: 2rem;
-        border: none;
-        outline: none;
-        text-align: start;
-        padding-left: 2rem;
-        cursor: pointer;
-        background-color: transparent;
-        border-bottom: 1px solid $light-gray;
-        border-left: none;
-        &:hover {
-            background-color: $light-gray;
-        }
-    }
-
-    .status-button-active {
-        background-color: $light-gray;
-        border-left: 3px solid $theme-primary;
-    }
-}
-
-.close {
-    cursor: pointer;
-    font-size: 1.5rem;
-}
